+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/pagination/impl/ListJobPaginator.kt
@@ -11,20 +11,20 @@ import kotlin.math.min
 class ListJobPaginator<out T, V : PaginatedJobMetadata, K> private constructor(
     override val jobData: T,
     override val jobMetadata: V,
-    private val pageExtractor: (pageSize: Long, jobData: T) -> List<K>,
+    private val pageExtractor: (pageSize: Int, jobData: T) -> List<K>,
 ) : JobPaginator<T, V, K> {
 
     data class Builder<T, V : PaginatedJobMetadata, K>(
         var jobData: T,
         var jobMetadata: V,
-        var pageExtractor: (pageSize: Long, jobData: T) -> List<K>,
+        var pageExtractor: (pageSize: Int, jobData: T) -> List<K>,
     ) {
 
         fun build() = ListJobPaginator(jobData, jobMetadata, pageExtractor)
 
     }
 
-    override val pageSize: Long = jobMetadata.pageSize
+    override val pageSize: Int = jobMetadata.pageSize
 
     private var pagesLeft = ceil(jobMetadata.maxCountPerExecution.toDouble() / pageSize).toLong()
     private var elementsLeft = jobMetadata.maxCountPerExecution
@@ -71,6 +71,6 @@ class ListJobPaginator<out T, V : PaginatedJobMetadata, K> private constructor(
 fun <T, V : PaginatedJobMetadata, K> listJobPaginator(
     jobData: T,
     jobMetadata: V,
-    pageExtractor: (pageSize: Long, jobData: T) -> List<K>,
+    pageExtractor: (pageSize: Int, jobData: T) -> List<K>,
     buildActions: (ListJobPaginator.Builder<T, V, K>.() -> Unit)? = null
 ) = ListJobPaginator.Builder(jobData, jobMetadata, pageExtractor).apply(buildActions ?: {}).build()
