+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/CommonOrderJobMetadata.kt
@@ -3,7 +3,7 @@ package io.scheduler.comparison.quartz.jobs.state
 data class CommonOrderJobMetadata(
     override val jobName: String,
     override val jobCron: String,
-    override val pageSize: Long,
-    override val maxCountPerExecution: Long,
+    override val pageSize: Int,
+    override val maxCountPerExecution: Int,
 ) : CronJobMetadata, PaginatedJobMetadata
