+++ scheduler-quartz/src/main/kotlin/io/scheduler/comparison/quartz/jobs/state/DedicatedOrderJobMetadata.kt
@@ -3,6 +3,6 @@ package io.scheduler.comparison.quartz.jobs.state
 data class DedicatedOrderJobMetadata(
     override val jobName: String,
     override val jobCron: String,
-    override val pageSize: Long,
-    override val maxCountPerExecution: Long,
+    override val pageSize: Int,
+    override val maxCountPerExecution: Int,
 ) : CronJobMetadata, PaginatedJobMetadata
